fixed issue in array test, renamed files.
