+++ com.ibm.wala.core/src/testSubjects11/java/privateInterfaceMethods/testArrayReturn/testArrayReturn.java
@@ -1,9 +1,12 @@
-package testArrayReturn;
+package privateInterfaceMethods.testArrayReturn;
+
+import privateInterfaceMethods.testArrayReturn.returnArray;
 
 public class testArrayReturn {
   public void main(String[] args){
     returnArray testee = new returnArray();
-    testee.GetT({1,2,3,4});
+    int[] arrayArg = {1,2,3,4};
+    testee.RetT(arrayArg);
     return;
   }
