+++ com.ibm.wala.core/src/testSubjects11/java/privateInterfaceMethods/testDoubleReturn/ReturnDouble.java
@@ -1,0 +1,7 @@
+package privateInterfaceMethods.testDoubleReturn;
+
+import privateInterfaceMethods.PrivateInterface;
+
+public class ReturnDouble implements PrivateInterface {
+
+}
