--- com.ibm.wala.core/src/testSubjects11/java/privateInterfaceMethods/testDoubleReturn/returnDouble.java
@@ -1,5 +1,0 @@
-package testArrayReturn;
-
-public class ReturnDouble implements PrivateInterface{
-
-}
