+++ com.ibm.wala.core/src/testSubjects11/java/privateInterfaceMethods/testIntReturn/testIntReturn.java
@@ -1,4 +1,4 @@
-package testIntReturn;
+package privateInterfaceMethods.testIntReturn;
 
 public class testIntReturn {
