+++ com.ibm.wala.core/src/testSubjects11/java/privateInterfaceMethods/testStringReturn/testStringReturn.java
@@ -1,4 +1,4 @@
-package testStringReturn;
+package privateInterfaceMethods.testStringReturn;
 
 public class testStringReturn {
   public static void main(String[] args){
